+++ apx_sdk/CHANGELOG.md
@@ -18,6 +18,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Renamed `JsonSigner` type to `VerificationMethod`.
 - Renamed `signer` field on `JsonSignatureData` type to `verification_method`.
 
+## Deprecated
+
+- Marked `sign_object_rsa` as deprecated.
+
 ## [0.7.0] - 2024-12-19
 
 ### Added